--- packages/mui-codemod/src/v6.0.0/theme-style-overrides/index.js
@@ -1,1 +1,0 @@
-export { default } from './theme-style-overrides';
