+++ packages/mui-codemod/src/v6.0.0/theme-v6/index.js
@@ -1,0 +1,1 @@
+export { default } from './theme-v6';
