+++ packages/mui-codemod/src/v6.0.0/theme-v6/theme-v6.js
